requirements.txt
python-telegram-bot==21.6
anthropic>=0.40.0
openai>=1.50.0
python-dotenv>=1.0.0
requests>=2.31.0
pydub>=0.25.1
Pillow>=10.0.0
fpdf2>=2.7.0
